+++ internal/agent/client.go
@@ -1,3 +1,8 @@
+// Copyright 2023 the Paw Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
 package agent
 
 import (
@@ -5,12 +10,15 @@ import (
 	"crypto"
 	"encoding/json"
 	"errors"
-	"net"
 	"time"
 
 	"golang.org/x/crypto/ssh"
 	sshagent "golang.org/x/crypto/ssh/agent"
 	"lucor.dev/paw/internal/paw"
+)
+
+const (
+	dialTimeout = 100 * time.Millisecond
 )
 
 type PawAgent interface {
@@ -47,7 +55,7 @@ type client struct {
 // NewClient returns a Paw agent client to manage sessions and SSH keys
 // The communication with agent is done using the SSH agent protocol.
 func NewClient(socketPath string) (PawAgent, error) {
-	a, err := net.Dial("unix", socketPath)
+	a, err := dialWithTimeout(socketPath, dialTimeout)
 	if err != nil {
 		return nil, err
 	}