+++ internal/paw/storage.go
@@ -1,3 +1,8 @@
+// Copyright 2021 the Paw Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
 package paw
 
 import (
@@ -5,6 +10,7 @@ import (
 	"fmt"
 	"io"
 	"path/filepath"
+	"runtime"
 )
 
 const (
@@ -12,7 +18,9 @@ const (
 	configFileName  = "config.json"
 	keyFileName     = "key.age"
 	vaultFileName   = "vault.age"
+	lockFileName    = "paw.lock"
 	socketFileName  = "agent.sock"
+	namedPipe       = `\\.\pipe\paw`
 )
 
 type Storage interface {
@@ -21,6 +29,7 @@ type Storage interface {
 	VaultStorage
 	ItemStorage
 	SocketAgentPath() string
+	LockFilePath() string
 }
 type ConfigStorage interface {
 	LoadConfig() (*Config, error)
@@ -80,7 +89,14 @@ func itemPath(s Storage, vaultName string, itemID string) string {
 }
 
 func socketAgentPath(s Storage) string {
+	if runtime.GOOS == "windows" {
+		return namedPipe
+	}
 	return filepath.Join(s.Root(), socketFileName)
+}
+
+func lockFilePath(s Storage) string {
+	return filepath.Join(s.Root(), lockFileName)
 }
 
 func encrypt(key *Key, w io.Writer, v interface{}) error {